CLP_Calculator/requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
python-dotenv==1.0.1
Flask-WTF==1.2.1
